requirements.txt
mysql-connector-python>=8.0.33
pytest>=7.4.0
pytest-cov>=4.1.0
